requirements.txt
flask
pynput
pywin32
pystray
pillow
pillow
google-genai
